+++ src/main/java/Logs/log.xml
@@ -1,18 +1,9 @@
 [TestNG] Running:
-  /home/rb/.IdeaIC14/system/temp-testng-customsuite.xml
+        /home/rb/.IdeaIC14/system/temp-testng-customsuite.xml
 
 
-===============================================
-Custom suite
-Total tests run: 1, Failures: 0, Skips: 0
-===============================================
-
-[TestNG] Running:
-  /home/rb/.IdeaIC14/system/temp-testng-customsuite.xml
-
-
-===============================================
-Custom suite
-Total tests run: 1, Failures: 0, Skips: 0
-===============================================
+        ===============================================
+        Custom suite
+        Total tests run: 1, Failures: 0, Skips: 0
+        ===============================================
