--- src/main/resources/basics.xml
@@ -1,22 +1,0 @@
-<suite name="Simple test" verbose="10">
-    <!--<parameter name="selenium.host" value="localhost" />-->
-    <!--<parameter name="selenium.port" value="3737" />-->
-    <!--<parameter name="selenium.browser" value="*firefox" />-->
-    <parameter name="base.url" value="http://demo.opensourcecms.com/wordpress/" />
-
-    <test name="Write new post" preserve-order="true">
-        <classes>
-            <class name="FirstTest">
-                <methods>
-                    <include name="launchSite" />
-                    <include name="openAdminPage" />
-                    <include name="loginAsAdmin" />
-                    <include name="navigateNewPost" />
-                    <include name="writeBlogPost" />
-                    <include name="publishBlogPost" />
-                    <include name="verifyBlogPost" />
-                </methods>
-            </class>
-        </classes>
-    </test>
-</suite>
